Fixes for c++20 (#1512)
It turns out that c++20 restricts the usage of forward
declarations of objects used in templates, particularly
std::vector. In this case functions defined inside the
class make use of an incomplete class - thanks to the
forward declaration -  when accessing functions of the
vector. The solutiion  in this case - since we have a
circular dependency here - is to move ALL function
definitions of ClosestPair2D  out of the class declaration
and after Point, so they always see a complete object.